# sqlescape: Improve performance of EscapeID