@mattlord 2022-03-08
Nice! I’m still a golang n00b so I’m always learning interesting new things — in this case the testing lib’s Benchmark support. I love it.

I was operating under the assumption that when not dictated by the interface you would use a byte buffer when you want a byte slice in the end and a string builder when you want a string. But I saw byte buffers being used that violated this rule and I wasn’t sure if it was just old code or there were other considerations (highly concurrent envs can sometimes turn assumptions on their head). @vmg could probably comment on that.