@mattrobenolt 2022-03-08
> I was operating under the assumption that when not dictated by the interface you would use a byte buffer when you want a byte slice in the end and a string builder when you want a string. But I saw byte buffers being used that violated this rule and I wasn’t sure if it was just old code or there were other considerations

Yup, this is generally true, but `strings.Builder` is much newer than `bytes.Buffer`. They are effectively the same thing, except as you mentioned, when you expect a `string` in the end, `strings.Builder` saves an allocation, whereas `bytes.Buffer` is better when you expect `[]byte` at the end. `strings.Builder` contains a nice trick to convert the underlying bytes to a string: https://github.com/golang/go/blob/master/src%2Fstrings%2Fbuilder.go#L47 :sunglasses: which avoids the extra allocation.

> highly concurrent envs can sometimes turn assumptions on their head

Neither `bytes.Buffer` or `strings.Builder` are safe for concurrent use, so we didn't change the assumption here.